Use separate header and source files for each class
Move IPFSOnlyUrlInterceptor and QMLRuntime classes into it's own header and source files
